+++ BlockBlizzard.lua
@@ -1,20 +1,20 @@
 Scorpio "AshToAsh.BlizzardSkin.BlockBlizzard" ""
 
+local hiddenFrame = Frame("BlockBlizzard")
+hiddenFrame:Hide()
+
 -- 不替换RegisterEvent，使用这种方式防止taint
 local function clearEvents(frame, event)
-    print("ClearEvents", frame:GetName(), event)
     frame:UnregisterAllEvents()
 end
 
 local function clearScript(frame, scriptType, handler)
-    print("ClearScript", frame:GetName(), scriptType, handler)
     if handler then
         frame:SetScript(scriptType, nil)
     end
 end
 
 local function clearFrame(frame)
-    print("ClearFrame", frame:GetName())
     hooksecurefunc(frame, "RegisterEvent", clearEvents)
     hooksecurefunc(frame, "RegisterUnitEvent", clearEvents)
     hooksecurefunc(frame, "RegisterAllEvents", clearEvents)
@@ -24,16 +24,17 @@ local function clearFrame(frame)
     frame:SetScript("OnUpdate", nil)
     frame:SetScript("OnEnter", nil)
     frame:SetScript("OnLeave", nil)
-    frame:SetScript("OnClick", nil)
+    if frame:HasScript("OnClick") then
+        frame:SetScript("OnClick", nil)
+    end
     frame:UnregisterAllEvents()
-    frame:Hide()
 end
 
-local function HideBlizzardFrames(frame)
+local function BlockBlizzardFrames(frame)
     if frame and frame:IsObjectType("Frame") then
         local children = { frame:GetChildren() }
         for _, child in ipairs(children) do
-           HideBlizzardFrames(child)
+           BlockBlizzardFrames(child)
         end
 
         clearFrame(frame)
@@ -46,21 +47,34 @@ end
 
 __Async__()
 function OnEnable()
-    BlockPartyFrames()
+    if CompactPartyFrame then
+        BlockUnitFrames(CompactPartyFrame)
+    end
+
+    if not IsAddOnLoaded('Blizzard_CompactRaidFrames') then
+        while NextEvent('ADDON_LOADED') ~= 'Blizzard_CompactRaidFrames' do end
+    end
+
+    BlockUnitFrames(CompactRaidFrameContainer)
 end
 
 __NoCombat__()
-function BlockPartyFrames()
-    HideBlizzardFrames(CompactPartyFrame)
-end
-
-__NoCombat__()
-__AddonSecureHook__('Blizzard_CompactRaidFrames', 'CompactRaidFrameContainer_OnLoad')
-function BlockRaidFrames(frame)
-    HideBlizzardFrames(frame)
+function BlockUnitFrames(frame)
+    frame:SetParent(hiddenFrame)
+    BlockBlizzardFrames(frame)
 end
 
 __SecureHook__()
 function CompactUnitFrame_OnLoad(frame)
-    HideBlizzardFrames(frame)
+    local name = frame:GetName()
+    if name and (name:find("CompactRaid") or name:find("CompactParty")) then
+        BlockBlizzardFrames(frame)
+    end
+end
+
+__SecureHook__()
+__NoCombat__()
+function CompactPartyFrame_Generate()
+    CompactPartyFrame:SetParent(hiddenFrame)
+    BlockBlizzardFrames(CompactPartyFrame)
 end